requirements.txt
# --- Core Framework ---
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.1
python-multipart>=0.0.6

# --- AWS ---
boto3>=1.34.11

# --- Utilities & Logging ---
jinja2>=3.1.2
python-json-logger>=2.0.7
structlog>=23.2.0
redis>=5.0.1
gunicorn>=21.2.0
sentry-sdk>=1.39.1
PyJWT>=2.8.0
certifi>=2023.11.17

# --- LangGraph and LangChain (v1.0+) ---
langgraph>=1.0.0
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.1  # Note: 1.0.0 not available yet, using latest 0.4.1
langchain-core>=1.0.0
langgraph-cli>=0.4.11  # Note: 1.0.0 not available yet, using latest 0.4.11
langsmith>=0.1.0

# --- MongoDB ---
pymongo>=4.9,<4.10
motor==3.6.0

# --- OpenAI ---
openai>=1.54.0

# --- Data & Scraping ---
beautifulsoup4>=4.12.3
lxml>=5.2.2
numpy>=1.26.4
ddgs>=1.0.0  # 최신 DuckDuckGo 검색 라이브러리 (권장)
duckduckgo-search>=5.1.0  # 구버전 호환성 (fallback)

# --- Data Validation ---
pydantic>=2.0.0